+++ src/Tecan/Rack/A.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+class A extends RackBase
+{
+    public function type(): string
+    {
+        return 'Eppis 24x0.5 ml Cooled';
+    }
+
+    public function name(): string
+    {
+        return 'A';
+    }
+}