+++ src/Tecan/Rack/DestPCR.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+class DestPCR extends RackBase
+{
+    public function type(): string
+    {
+        return '96 Well PCR ABI semi-skirted';
+    }
+
+    public function name(): string
+    {
+        return 'DestPCR';
+    }
+}