+++ src/Tecan/Rack/MM.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+class MM extends RackBase
+{
+    public function type(): string
+    {
+        return 'Eppis 32x1.5 ml Cooled';
+    }
+
+    public function name(): string
+    {
+        return 'MM';
+    }
+}