+++ src/Tecan/Rack/MPCDNA.php
@@ -1,0 +1,16 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+class MPCDNA extends RackBase
+{
+    public function type(): string
+    {
+        return 'MP cDNA';
+    }
+
+    public function name(): string
+    {
+        return 'MPCDNA';
+    }
+}