+++ CHANGELOG.md
@@ -13,6 +13,8 @@ See [GitHub releases](https://github.com/mll-lab/php-utils/releases).
 
 ### Changed
 
+- add specific class for each `MLLLabWareRack`-type
+- use `MLLLabWareRack` as factory for the `Rack`-objects
 - Breaking Change: delete `CustomRack`-class
 
 ## v3.2.0